SECURITY ISSUE: Removing crt files that were checked in with private keys intact.  Private keys should never be checked into subversion.
I recognize that removing these files does not actually delete them from svn, 
but it does make them less obvious.  At this point, we have completely replaced
the old CA with a new testing CA, and so these certificates should not be
trusted on any hosts anyways.  But removing them from this directory will make
any software that depends on them being here break, and therefore make it
obvious which software needs to be updated to not depend on insecure storage of
key files.  Removing them will make it easier to track down this software.



git-svn-id: file:///var/code/dataone/dataone/software/cicore/trunk/d1_integration@9208 a8e578cc-b067-0410-9529-a28e5a6d5a41